Move profiling module to modules.utils package.
Also update changelog.
